fix build

+++ leptonica/src/utils.c
@@ -3459,7 +3459,7 @@ l_uint64  limit, ratio;
         return 0;
     }
 
-    limit = sqrt(n);
+    limit = sqrt((double)n);
     for (div = 3; div < limit; div += 2) {
        ratio = n / div;
        if (ratio * div == n) {
